requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
blinker==1.4
Brotli==1.0.9
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
Django==5.0.6
django-jazzmin==3.0.0
fasteners==0.14.1
frozenlist==1.4.1
future==0.18.2
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
mutagen==1.45.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pycryptodomex==3.11.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-dotenv==1.0.1
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
redis==5.0.4
redis-dict==2.5.0
reportlab==3.6.8
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.2
wadllib==1.3.6
websockets==9.1
xdg==5.0.0
yarl==1.9.4
yt-dlp==2022.4.8
zipp==1.0.0
